Support building with mingw
This mostly covers our bases, just leaving libjpeg-turbo as the perpetual issue that it is, and more than likely still requiring some build-related config to ensure everything is statically linked still (aside from DX9).

+++ src/Client/WarFare/UITransactionDlg.cpp
@@ -346,7 +346,7 @@ void CUITransactionDlg::EnterTransactionState()
 			continue;
 		}
 
-		if (pItemExt->dwID != iExt)
+		if (pItemExt->dwID != static_cast<uint32_t>(iExt))
 			continue;
 
 		e_PartPosition ePart = PART_POS_UNKNOWN;
